ta_nemo_guardrails/pip_requirements.txt
torch==2.4.1
torchdiffeq==0.2.4
torchmetrics==1.4.2
torchsde==0.2.6
torchvision==0.19.1
absl-py==2.1.0
accelerate==0.34.2
accelerated-scan==0.2.0
addict
aiohappyeyeballs
aiohttp
aiosignal
alabaster
aniso8601
annotated-types
annoy
antlr4-python3-runtime
anyio
appdirs
arrow
asciitree
asgiref
asttokens
attrdict
attrs
audioread
babel
beautifulsoup4
black
blinker
boto3==1.35.27
botocore==1.35.27
braceexpand==0.1.7
Brotli==1.0.9
cachetools==5.3.0
canvasapi==3.3.0
cdifflib==1.2.6
certifi==2023.11.17
cffi==1.15.0
chardet==5.2.0
charset-normalizer
click==8.0.2
clip==0.2.0
cloudpickle==3.0.0
colorama
coloredlogs
comm
contourpy
cryptography
cycler==0.12.1
Cython==0.29.36
cytoolz==0.12.3
dataclasses-json==0.6.7
DataProperty==1.0.1
datasets==3.0.0
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
diffusers==0.30.3
dill
Distance
distro
Django
django-auditlog
django-filter
django-js-asset
django-mptt
djangorestframework
docker-pycreds
docopt
docutils
drf-excel
drf-flex-fields
editdistance
einops
einops-exts==0.0.4
et-xmlfile==1.1.0
evaluate==0.4.3
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.8.0.post1
fastapi==0.115.4
fastembed @ git+https://github.com/qdrant/fastembed.git@main#egg=fastembed
fasteners==0.19
fasttext==0.9.3
fiddle==0.3.0
filelock==3.12.0
Flask==2.2.5
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.3.3
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
g2p-en==2.1.0
gdown==5.2.0
git-lfs==1.6
gitdb==4.0.11
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
graphviz
greenlet
grpcio==1.66.1
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==0.17.1
httplib2==0.22.0
httpx
huggingface-hub>=0.24
humanfriendly
hydra-core
hyperframe
idna
ijson
imageio
imagesize
importlib_metadata
inflect
iniconfig
intervaltree
ipykernel
ipython
ipywidgets
isort
itsdangerous
Janome
jedi
jieba
Jinja2
jiter
jiwer==2.5.2
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.4.0
jupyter_core==5.3.0
jupyterlab_widgets==3.0.13
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.43
langchain-nvidia-ai-endpoints==0.2.2
langchain-openai==0.1.22
langchain-text-splitters==0.2.4
langsmith==0.1.138
lark==1.1.9
latexcodec==3.0.0
lazy_loader==0.4
ldap3==2.9.1
Levenshtein==0.22.0
lhotse==1.27.0
libcst==1.4.0
librosa==0.10.2.post1
lightning-utilities==0.11.7
lilcom==1.8.0
llvmlite==0.43.0
git+https://github.com/Elan456/lm-evaluation-harness.git@f47213aa0a63593a0f683d8654302ca5a02d9d7e#egg=lm_eval
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mdurl==0.1.2
megatron-energon==2.3.0
megatron_core==0.5.0
mmh3==4.1.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
NEMO==6.0.3
nemo_text_processing==1.1.0
nemo_toolkit==1.23.0
nemoguardrails==0.10.1
nerfacc==0.5.3
nest_asyncio==1.5.6
networkx==3.3
nltk==3.9.1
numba==0.60.0
numcodecs==0.13.0
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf
onnx
onnxruntime
open-clip-torch
openai
OpenCC
openpyxl
orjson
packaging
pandas
pangu
parameterized
parso
pathspec
pathvalidate
peft
pesq
pexpect
pickleshare==0.7.5
pillow==10.3.0
plac==1.4.3
platformdirs==3.2.0
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
progress==1.6
prompt-toolkit
proto-plus
protobuf
psutil
ptyprocess
pure_eval==0.2.0
py_rust_stemmers==0.1.3
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1_modules==0.2.8
pybind11
pybtex
pybtex-docutils
pycparser
pydantic
pydantic_core
pydub
Pygments
pyloudnorm
PyMCubes
pymodbus
pynini
pyOpenSSL
pyparsing
pypdfium2
pypinyin
pypinyin-dict
PySide6
PySide6_Addons
PySide6_Essentials
PySocks
pystoi
pytablewriter
pytest
pytest-mock
pytest-runner
python-dateutil
python-docx
python-pptx
pytorch-lightning
pytz
pyu2f
PyYAML
pyzmq
questionary
rapidfuzz
regex
requests
requests-toolbelt
resampy
retrying
rich
rouge_score
rsa
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==0.4.2
s3transfer==0.10.2
sacrebleu==2.4.3
sacremoses==0.1.1
safetensors
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.1
sentencepiece==0.2.0
sentry-sdk
setproctitle==1.3.3
shellingham==1.5.4
shiboken6==6.8.0.2
simpleeval==1.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
sox==1.5.0
soxr==0.5.0.post1
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.3
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
sqlitedict==2.1.0
sqlparse==0.5.1
stack-data==0.6.2
starlette==0.41.2
sympy==1.13.3
tabledata==1.3.3
tabulate==0.9.0
taming-transformers==0.0.1
tcolorpy==0.1.6
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorstore==0.1.45
termcolor==2.4.0
text-unidecode==1.3
textdistance==4.6.3
texterrors==0.5.1
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==0.6.13
tokenizers==0.20.0
toml==0.10.2
toolz==0.12.1
tornado==6.3.2
tqdm
tqdm-multiprocess
traitlets
trampoline
transformers>=4.45.1
trimesh
triton
typed-ast
typeguard
typepy
typer
types-python-dateutil
typing-inspect
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
wandb
watchdog
wcwidth
webdataset
Werkzeug
wget==3.2
widgetsnbextension==4.0.13
word2number==1.1
wrapt==1.16.0
XlsxWriter==3.2.0
xxhash==3.2.0
yarl==1.8.2
youtokentome==1.0.6
zarr==2.18.3
zipp==3.15.0
zstandard==0.23.0
rank_bm25